requirements.txt
streamlit
Pillow==10.3.0
openai==1.37.0
python-dotenv==1.0.1